fix: wrap browse page in Suspense and move useSearchParams to client component
